big chunk of order processing is completed.
To be able to get selections from one page to another is really challenging if you're not using any serverside scripts however , there are futures built-in browsers such as Local Storage , Session Storage or Cookies to be able to direct values. Since we're not using back-end in this project , I choosed to use localStorage because it has most amount of data capacity to hold and easy to clean afterwards. But of course using back-end is always the safest option.